+++ doc/release/release_notes.rst
@@ -17,6 +17,12 @@ Release Notes
         Added example on how to use `SecurityRole` and `UserSecurityRole` to
         manage security roles for users.
 
+    .. change:: changed
+        :tags: documentation
+
+        Added examples to show how to list a user's assigned tasks and all
+        users assigned to a task. :ref:`example/assignments_and_allocations`.
+
   .. change:: new
     :tags: server location, thumbnail