+++ doc/release/release_notes.rst
@@ -23,8 +23,8 @@ Release Notes
         Added examples to show how to list a user's assigned tasks and all
         users assigned to a task. :ref:`example/assignments_and_allocations`.
 
-  .. change:: new
-    :tags: server location, thumbnail
+    .. change:: new
+        :tags: server location, thumbnail
 
         Added method :meth:`get_thumbnail_url() <ftrack_api.entity.location.ServerLocationMixin.get_thumbnail_url>`
         to server location, which can be used to retrieve a thumbnail URL.