+++ doc/release/release_notes.rst
@@ -14,6 +14,12 @@ Release Notes
     .. change:: new
         :tags: documentation
 
+        Added :ref:`example <example/entity_links>` on how to manage entity
+        links from the API.
+
+    .. change:: new
+        :tags: documentation
+
         Added :ref:`example <example/manage_custom_attribute_configuration>` on
         how to manage custom attribute configurations from the API.